requirements.txt
numpy
fake-bpy-module-4.0